Commit recent progress
I am going to replicate the current site in a temporary app so I can get the
site running on django. I did a mass commit of the project when it was in a
functional state, and I will create a tagged version that will be used until the
full site is ready. This will force me to take care of the necessary bits to get
the full application running live. None of the code in the 'diesel' app will be
used, since an error in one of those models could prevent the site from working
even if the code isn't used.